FIX: fixed various issues in findgoal env and marlgrid base env

+++ MAWM/envs/marl_grid/agents.py
@@ -163,12 +163,12 @@ class GridAgentInterface(GridAgent):
                 comm_space = gym.spaces.Box(low=0.0,
                                             high=comm_dim,
                                             shape=(n_agents, comm_len),
-                                            dtype=np.float32)
+                                            dtype=float)
                 comm_act_space = gym.spaces.Box(low=0.0,
                                                 high=comm_dim,
                                                 shape=(comm_len,),
-                                                dtype=np.float32)
-                self.comm = np.zeros((comm_len,), dtype=np.float32)
+                                                dtype=float)
+                self.comm = np.zeros((comm_len,), dtype=float)
         else:
             self.observe_comm = False
 
@@ -190,17 +190,17 @@ class GridAgentInterface(GridAgent):
             if self.observe_rewards:
                 obs_space['reward'] = gym.spaces.Box(low=-np.inf, high=np.inf,
                                                      shape=(),
-                                                     dtype=np.float32)
+                                                     dtype=float)
             if self.observe_position:
                 # discrete position space is handled in base.py
                 obs_space['position'] = gym.spaces.Box(low=0, high=1,
                                                        shape=(n_agents, 2),
-                                                       dtype=np.float32)
+                                                       dtype=float)
             if self.observe_self_position:
                 # discrete position space is handled in base.py
                 obs_space['selfpos'] = gym.spaces.Box(low=0, high=1,
                                                       shape=(2,),
-                                                      dtype=np.float32)
+                                                      dtype=float)
             if self.observe_done:
                 obs_space['done'] = gym.spaces.Discrete(n=2)
             if self.observe_self_env_act:
@@ -210,7 +210,7 @@ class GridAgentInterface(GridAgent):
             if self.observe_t:
                 obs_space['t'] = gym.spaces.Box(low=-np.inf, high=np.inf,
                                                 shape=(),
-                                                dtype=np.float32)
+                                                dtype=float)
             if self.observe_identity:
                 obs_space['identity'] = gym.spaces.Discrete(n=2)
             if self.observe_comm:
@@ -221,17 +221,17 @@ class GridAgentInterface(GridAgent):
             if self.observe_rewards:
                 obs_space += gym.spaces.Box(low=-np.inf, high=np.inf,
                                             shape=(),
-                                            dtype=np.float32)
+                                            dtype=float)
             if self.observe_position:
                 # discrete position space is handled in base.py
                 obs_space += gym.spaces.Box(low=0, high=1,
                                             shape=(n_agents, 2),
-                                            dtype=np.float32)
+                                            dtype=float)
             if self.observe_self_position:
                 # discrete position space is handled in base.py
                 obs_space += gym.spaces.Box(low=0, high=1,
                                             shape=(2,),
-                                            dtype=np.float32)
+                                            dtype=float)
             if self.observe_done:
                 obs_space += gym.spaces.Discrete(n=2)
             if self.observe_self_env_act:
@@ -241,7 +241,7 @@ class GridAgentInterface(GridAgent):
             if self.observe_t:
                 obs_space += gym.spaces.Box(low=-np.inf, high=np.inf,
                                             shape=(),
-                                            dtype=np.float32)
+                                            dtype=float)
             if self.observe_identity:
                 obs_space += gym.spaces.Discrete(n=2)
             if self.observe_comm:
